tests/wpt/tests/css/css-text/parsing/hyphens-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Text Module Test: parsing hyphens with valid values</title>
<link rel="help" href="https://www.w3.org/TR/css-text-3/#propdef-hyphens">
<meta name="assert" content="hyphens supports the full grammar 'none | manual | auto'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value("hyphens", "none");
test_valid_value("hyphens", "manual");
test_valid_value("hyphens", "auto");
</script>
</body>
</html>
